Move Fitter to separate files.
